requirements.txt
streamlit == 1.31.0
pandas == 2.1.4
joblib == 1.3.2
nltk == 3.8.1
sastrawi == 1.0.1
scikit-learn == 1.2.2
openpyxl == 3.1.2
tweet-preprocessor == 0.6.0